+++ components/student/LpsciIndividualStudent/LpsciIndividualStudentProfile/LpsciIndividualStudentSecondary/LpsciIndividualStudentStatus/LpsciIndividualStudentStatusCards/LpsciIndividualStudentStatusCard/index.tsx
