+++ components/student/LpsciIndividualStudent/LpsciIndividualStudentProfile/LpsciIndividualStudentSecondary/LpsciIndividualStudentStatus/LpsciIndividualStudentStatusCards/LpsciIndividualStudentStatusCard/styles.ts
